Factor volume-data IO checks into helper function
