+++ tests/Unit/IO/H5/Test_VolumeData.cpp
@@ -3,8 +3,6 @@
 
 #include "Framework/TestingFramework.hpp"
 
-#include <algorithm>
-#include <boost/iterator/transform_iterator.hpp>
 #include <cstddef>
 #include <cstdint>
 #include <memory>
@@ -13,150 +11,16 @@
 
 #include "DataStructures/DataVector.hpp"
 #include "DataStructures/Tensor/TensorData.hpp"
+#include "Helpers/IO/VolumeDataHelpers.hpp"
 #include "IO/H5/AccessType.hpp"
 #include "IO/H5/File.hpp"
 #include "IO/H5/VolumeData.hpp"
 #include "NumericalAlgorithms/Spectral/Spectral.hpp"
 #include "NumericalAlgorithms/SphericalHarmonics/Strahlkorper.hpp"
 #include "NumericalAlgorithms/SphericalHarmonics/YlmSpherepack.hpp"
-#include "Utilities/Algorithm.hpp"
-#include "Utilities/ErrorHandling/Error.hpp"
 #include "Utilities/FileSystem.hpp"
-#include "Utilities/MakeString.hpp"
-#include "Utilities/Numeric.hpp"
 
 namespace {
-template <typename T>
-T multiply(const double obs_value, const T& component) {
-  T result = component;
-  for (auto& t : result) {
-    t *= obs_value;
-  }
-  return result;
-}
-
-// Check that the volume data is correct
-template <typename DataType>
-void check_volume_data(
-    const std::string& h5_file_name, const uint32_t version_number,
-    const size_t observation_id, const double observation_value,
-    const std::vector<DataType>& tensor_components_and_coords,
-    const std::vector<std::string>& grid_names,
-    const std::vector<std::vector<Spectral::Basis>>& bases,
-    const std::vector<std::vector<Spectral::Quadrature>>& quadratures,
-    const std::vector<std::vector<size_t>>& extents,
-    const std::vector<std::string>& expected_components,
-    const std::vector<std::vector<size_t>>& grid_data_orders) {
-  h5::H5File<h5::AccessType::ReadOnly> file_read{h5_file_name};
-  const auto& volume_file =
-      file_read.get<h5::VolumeData>("/element_data", version_number);
-
-  CHECK(extents == volume_file.get_extents(observation_id));
-  CHECK(volume_file.get_observation_value(observation_id) == observation_value);
-  // Check that all of the grid names were written correctly by checking their
-  // equality of elements
-  const std::vector<std::string> read_grid_names =
-      volume_file.get_grid_names(observation_id);
-  [&read_grid_names, &grid_names]() {
-    auto sortable_grid_names = grid_names;
-    auto sortable_read_grid_names = read_grid_names;
-    std::sort(sortable_grid_names.begin(), sortable_grid_names.end(),
-              std::less<>{});
-    std::sort(sortable_read_grid_names.begin(), sortable_read_grid_names.end(),
-              std::less<>{});
-    REQUIRE(sortable_read_grid_names == sortable_grid_names);
-  }();
-  // Find the order the grids were written in
-  std::vector<size_t> grid_positions(read_grid_names.size());
-  for (size_t i = 0; i < grid_positions.size(); i++) {
-    auto grid_name = grid_names[i];
-    auto position =
-        std::find(read_grid_names.begin(), read_grid_names.end(), grid_name);
-    // We know the grid name is in the read_grid_names because of the previous
-    // so we know `position` is an actual pointer to an element
-    grid_positions[i] =
-        static_cast<size_t>(std::distance(read_grid_names.begin(), position));
-  }
-  auto read_bases = volume_file.get_bases(observation_id);
-  alg::sort(read_bases, std::less<>{});
-  auto read_quadratures = volume_file.get_quadratures(observation_id);
-  alg::sort(read_quadratures, std::less<>{});
-  // We need non-const bases and quadratures in order to sort them, and we
-  // need them in their string form,
-  const auto& stringify = [](const auto& bases_or_quadratures) {
-    std::vector<std::vector<std::string>> local_target_data{};
-    local_target_data.reserve(bases_or_quadratures.size() + 1);
-    for (const auto& element_data : bases_or_quadratures) {
-      std::vector<std::string> target_axis_data{};
-      target_axis_data.reserve(element_data.size() + 1);
-      for (const auto& axis_datum : element_data) {
-        target_axis_data.emplace_back(MakeString{} << axis_datum);
-      }
-      local_target_data.push_back(target_axis_data);
-    }
-    return local_target_data;
-  };
-  auto target_bases = stringify(bases);
-  alg::sort(target_bases, std::less<>{});
-  auto target_quadratures = stringify(quadratures);
-  alg::sort(target_quadratures, std::less<>{});
-  CHECK(target_bases == read_bases);
-  CHECK(target_quadratures == read_quadratures);
-
-  const auto read_components =
-      volume_file.list_tensor_components(observation_id);
-  CHECK(alg::all_of(read_components,
-                    [&expected_components](const std::string& id) {
-                      return alg::found(expected_components, id);
-                    }));
-  // Helper Function to get number of points on a particular grid
-  const auto accumulate_extents = [](const std::vector<size_t>& grid_extents) {
-    return alg::accumulate(grid_extents, 1, std::multiplies<>{});
-  };
-
-  const auto read_extents = volume_file.get_extents(observation_id);
-  std::vector<size_t> element_num_points(
-      boost::make_transform_iterator(read_extents.begin(), accumulate_extents),
-      boost::make_transform_iterator(read_extents.end(), accumulate_extents));
-  const auto read_points_by_element = [&element_num_points]() {
-    std::vector<size_t> read_points(element_num_points.size());
-    read_points[0] = 0;
-    for (size_t index = 1; index < element_num_points.size(); index++) {
-      read_points[index] =
-          read_points[index - 1] + element_num_points[index - 1];
-    }
-    return read_points;
-  }();
-  // Given a DataType, corresponding to contiguous data read out of a
-  // file, find the data which was written by the grid whose extents are
-  // found at position `grid_index` in the vector of extents.
-  const auto get_grid_data = [&element_num_points, &read_points_by_element](
-                                 const DataVector& all_data,
-                                 const size_t grid_index) {
-    DataType result(element_num_points[grid_index]);
-    // clang-tidy: do not use pointer arithmetic
-    std::copy(&all_data[read_points_by_element[grid_index]],
-              &all_data[read_points_by_element[grid_index]] +  // NOLINT
-                  element_num_points[grid_index],
-              result.begin());
-    return result;
-  };
-  // The tensor components can be written in any order to the file, we loop
-  // over the expected components rather than the read components because they
-  // are in a particular order.
-  for (size_t i = 0; i < expected_components.size(); i++) {
-    const auto& component = expected_components[i];
-    // for each grid
-    for (size_t j = 0; j < grid_names.size(); j++) {
-      CHECK(get_grid_data(
-                volume_file.get_tensor_component(observation_id, component),
-                grid_positions[j]) ==
-            multiply(observation_value,
-                     tensor_components_and_coords[grid_data_orders[j][i]]));
-    }
-  }
-}
-
 void test_strahlkorper() {
   constexpr size_t l_max = 12;
   constexpr size_t m_max = 12;
@@ -216,12 +80,13 @@ void test_strahlkorper() {
           observation_values[0]);
   }
 
-  check_volume_data(h5_file_name, version_number, observation_ids[0],
-                    observation_values[0], tensor_and_coord_data, {{grid_name}},
-                    {bases}, {quadratures}, {extents},
-                    {"InertialCoordinates_x", "InertialCoordinates_y",
-                     "InertialCoordinates_z", "TestScalar"},
-                    {{0, 1, 2, 3}});
+  TestHelpers::io::VolumeData::check_volume_data(
+      h5_file_name, version_number, "element_data"s, observation_ids[0],
+      observation_values[0], tensor_and_coord_data, {{grid_name}}, {bases},
+      {quadratures}, {extents},
+      {"InertialCoordinates_x", "InertialCoordinates_y",
+       "InertialCoordinates_z", "TestScalar"},
+      {{0, 1, 2, 3}}, {}, observation_values[0]);
 
   if (file_system::check_if_file_exists(h5_file_name)) {
     file_system::rm(h5_file_name, true);
@@ -266,52 +131,66 @@ void test() {
           observation_id, observation_value,
           std::vector<ElementVolumeData>{
               {{2, 2, 2},
-               {TensorComponent{first_grid + "/S",
-                                multiply(observation_value,
-                                         tensor_components_and_coords[0])},
-                TensorComponent{first_grid + "/x-coord",
-                                multiply(observation_value,
-                                         tensor_components_and_coords[1])},
-                TensorComponent{first_grid + "/y-coord",
-                                multiply(observation_value,
-                                         tensor_components_and_coords[2])},
-                TensorComponent{first_grid + "/z-coord",
-                                multiply(observation_value,
-                                         tensor_components_and_coords[3])},
-                TensorComponent{first_grid + "/T_x",
-                                multiply(observation_value,
-                                         tensor_components_and_coords[4])},
-                TensorComponent{first_grid + "/T_y",
-                                multiply(observation_value,
-                                         tensor_components_and_coords[5])},
-                TensorComponent{first_grid + "/T_z",
-                                multiply(observation_value,
-                                         tensor_components_and_coords[6])}},
+               {TensorComponent{
+                    first_grid + "/S",
+                    TestHelpers::io::VolumeData::multiply(
+                        observation_value, tensor_components_and_coords[0])},
+                TensorComponent{
+                    first_grid + "/x-coord",
+                    TestHelpers::io::VolumeData::multiply(
+                        observation_value, tensor_components_and_coords[1])},
+                TensorComponent{
+                    first_grid + "/y-coord",
+                    TestHelpers::io::VolumeData::multiply(
+                        observation_value, tensor_components_and_coords[2])},
+                TensorComponent{
+                    first_grid + "/z-coord",
+                    TestHelpers::io::VolumeData::multiply(
+                        observation_value, tensor_components_and_coords[3])},
+                TensorComponent{
+                    first_grid + "/T_x",
+                    TestHelpers::io::VolumeData::multiply(
+                        observation_value, tensor_components_and_coords[4])},
+                TensorComponent{
+                    first_grid + "/T_y",
+                    TestHelpers::io::VolumeData::multiply(
+                        observation_value, tensor_components_and_coords[5])},
+                TensorComponent{
+                    first_grid + "/T_z",
+                    TestHelpers::io::VolumeData::multiply(
+                        observation_value, tensor_components_and_coords[6])}},
                bases.front(),
                quadratures.front()},
               // Second Element Data
               {{2, 2, 2},
-               {TensorComponent{last_grid + "/S",
-                                multiply(observation_value,
-                                         tensor_components_and_coords[1])},
-                TensorComponent{last_grid + "/x-coord",
-                                multiply(observation_value,
-                                         tensor_components_and_coords[0])},
-                TensorComponent{last_grid + "/y-coord",
-                                multiply(observation_value,
-                                         tensor_components_and_coords[5])},
-                TensorComponent{last_grid + "/z-coord",
-                                multiply(observation_value,
-                                         tensor_components_and_coords[3])},
-                TensorComponent{last_grid + "/T_x",
-                                multiply(observation_value,
-                                         tensor_components_and_coords[6])},
-                TensorComponent{last_grid + "/T_y",
-                                multiply(observation_value,
-                                         tensor_components_and_coords[4])},
-                TensorComponent{last_grid + "/T_z",
-                                multiply(observation_value,
-                                         tensor_components_and_coords[2])}},
+               {TensorComponent{
+                    last_grid + "/S",
+                    TestHelpers::io::VolumeData::multiply(
+                        observation_value, tensor_components_and_coords[1])},
+                TensorComponent{
+                    last_grid + "/x-coord",
+                    TestHelpers::io::VolumeData::multiply(
+                        observation_value, tensor_components_and_coords[0])},
+                TensorComponent{
+                    last_grid + "/y-coord",
+                    TestHelpers::io::VolumeData::multiply(
+                        observation_value, tensor_components_and_coords[5])},
+                TensorComponent{
+                    last_grid + "/z-coord",
+                    TestHelpers::io::VolumeData::multiply(
+                        observation_value, tensor_components_and_coords[3])},
+                TensorComponent{
+                    last_grid + "/T_x",
+                    TestHelpers::io::VolumeData::multiply(
+                        observation_value, tensor_components_and_coords[6])},
+                TensorComponent{
+                    last_grid + "/T_y",
+                    TestHelpers::io::VolumeData::multiply(
+                        observation_value, tensor_components_and_coords[4])},
+                TensorComponent{
+                    last_grid + "/T_z",
+                    TestHelpers::io::VolumeData::multiply(
+                        observation_value, tensor_components_and_coords[2])}},
                bases.back(),
                quadratures.back()}});
     };
@@ -340,12 +219,13 @@ void test() {
   }
 
   for (size_t i = 0; i < observation_ids.size(); ++i) {
-    check_volume_data(
-        h5_file_name, version_number, observation_ids[i], observation_values[i],
-        tensor_components_and_coords, grid_names, bases, quadratures,
-        {{2, 2, 2}, {2, 2, 2}},
+    TestHelpers::io::VolumeData::check_volume_data(
+        h5_file_name, version_number, "element_data"s, observation_ids[i],
+        observation_values[i], tensor_components_and_coords, grid_names, bases,
+        quadratures, {{2, 2, 2}, {2, 2, 2}},
         {"S", "x-coord", "y-coord", "z-coord", "T_x", "T_y", "T_z"},
-        {{0, 1, 2, 3, 4, 5, 6}, {1, 0, 5, 3, 6, 4, 2}});
+        {{0, 1, 2, 3, 4, 5, 6}, {1, 0, 5, 3, 6, 4, 2}}, {},
+        observation_values[i]);
   }
 
   {
